@@ -1,6 +1,6 @@
 {
   "name": "suidouble-bot-score",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "private": true,
   "scripts": {
     "test": "tap -j=1 ./test/*.test.js",
@@ -16,51 +16,23 @@
     "node": "18.16.0"
   },
   "dependencies": {
-    "@chrisheanan/vue-gauge": "^1.2.2",
-    "@fastify/cookie": "^7.0.0",
-    "@fastify/cors": "^8.0.0",
-    "@fastify/formbody": "^7.0.1",
-    "@fastify/static": "^6.4.0",
-    "axios": "^0.23.0",
-    "fastify": "^4.2.0",
-    "fastify-mongoose-api": "^1.2.17",
-    "jose": "^4.8.3",
-    "lockfile": "^1.0.4",
-    "minio": "^7.0.28",
-    "mongoose": "^6.0.12",
-    "node-fetch": "^2.6.7",
-    "nodemailer": "^6.7.5",
-    "simple-statistics": "^7.8.3",
-    "suidouble": "^0.0.34",
-    "vue-sui": "^0.0.34",
-    "vue3-apexcharts": "^1.4.4",
-    "winston": "^2.3.1"
+    "@fastify/static": "^7.0.4",
+    "fastify": "^4.27.0"
   },
   "devDependencies": {
-    "@quasar/extras": "^1.0.0",
-    "@vue/cli-plugin-babel": "~5.0.8",
-    "@vue/cli-plugin-eslint": "~5.0.8",
-    "@vue/cli-service": "~5.0.8",
-    "@vue/compiler-sfc": "^3.0.0",
-    "apexcharts": "^3.35.3",
-    "awesome-qr": "^2.1.5-rc.0",
-    "babel-eslint": "^10.1.0",
-    "compression-webpack-plugin": "6.0.5",
-    "core-js": "^3.6.5",
-    "eslint": "^7.5.0",
-    "eslint-plugin-vue": "^7.0.0",
-    "node-polyfill-webpack-plugin": "^2.0.1",
-    "nodemon": "^2.0.14",
-    "pinia": "^2.0.14",
-    "quasar": "^2.11.6",
-    "tap": "^16.3.4",
-    "ts-loader": "^8.0.16",
-    "val-loader": "^2.1.2",
-    "vue": "^3.0.0",
-    "vue-cli-plugin-quasar": "~5.0.2",
-    "vue-recaptcha-v3": "^2.0.1",
-    "vue-router": "^4.0.11",
-    "worker-loader": "^3.0.8"
+    "@quasar/extras": "^1.16.11",
+    "@quasar/vite-plugin": "^1.7.0",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "pinia": "^2.1.7",
+    "quasar": "^2.16.4",
+    "simple-statistics": "^7.8.3",
+    "suidouble": "1.0.4-7",
+    "tap": "^16.3.10",
+    "vite": "^5.2.12",
+    "vue": "^3.4.27",
+    "vue-router": "^4.3.2",
+    "vue-sui": "1.0.4-10",
+    "vue3-apexcharts": "^1.5.3"
   },
   "eslintConfig": {
     "root": true,
