@@ -1,22 +1,9 @@
 const Fastify = require('fastify');
 
-// const fastifyServerAuth = require('./FastifyServerAuth.js');
-const fastifyCookie = require('@fastify/cookie');
-const fastifyFormbody = require('@fastify/formbody');
-// const fastifyMongooseAPI = require('fastify-mongoose-api');
-// const fastifyCors = require('@fastify/cors');
-
 class Server {
     constructor(params = {}) {
         this._logger = params.logger || null;
         this._port = params.server?.port || 8080;
-
-        this.db = params.db || null;
-
-        // if (!this.db) {
-        //     throw new Error('db param required for the server');
-        // }
-
         this._server = null;
 	}
 
@@ -29,29 +16,7 @@ class Server {
     }
 
     async beforeInit(fastify) {
-        fastify.register(fastifyCookie);
-        fastify.register(fastifyFormbody);
-        // fastify.register(fastifyCors, {
-        //     exposedHeaders: 'set-cookie',
-        //     credentials: true,
-        // });
-        // fastify.register(fastifyServerAuth, {
-        //         getUserByAuthCode: async (authCode)=>{
-        //             return await this.db.User.byAuthCode(authCode);
-        //         },
-        //     });
-
-        // fastify.register(fastifyMongooseAPI, {
-        //         models: this.db.connection.models,
-        //         checkAuth: (request)=>{
-        //             request.requireAuth();
-        //         },
-        //         prefix: '/api/',
-        //         setDefaults: true,
-        //         exposeVersionKey: false,
-        //         exposeModelName: true,
-        //         methods: ['list', 'get', 'post', 'patch', 'put', 'delete', 'options']
-        //     });
+        this.log('beforeInit...');
     }
 
     async init(beforeInit) {
