@@ -1,16 +1,15 @@
-const files = require.context('.', true, /\.vue$/)
-const modules = {}
+const files = import.meta.glob('./*.vue', { eager: true });
 
+const modules = {};
 
-files.keys().forEach((key) => {
-	if (key === './index.js') {
-		return;
-	}
-	modules[key.replace(/(\.\/|\.vue)/g, '')] = files(key);
-});
+for (const key in files) {
+	modules[key.replace(/(\.\/|\.vue)/g, '')] = files[key];
+}
 
 const routes = [];
 for (let key in modules) {
+	console.log(modules[key]);
+
 	routes.push({
 		path: modules[key].default.path,
 		component: modules[key].default,
